Moved utility functions from views.py to utils.py
