+++ hx_lti_initializer/views.py
@@ -6,7 +6,6 @@ It will set up the tool provider, create/retrive the user and pass along any
 other information that will be rendered to the access/init screen to the user.
 """
 
-from ims_lti_py.tool_provider import DjangoToolProvider
 from django.http import HttpResponseRedirect, HttpResponse
 from django.template import RequestContext
 
@@ -25,7 +24,7 @@ from hx_lti_initializer.models import LTIProfile, LTICourse, User
 from hx_lti_assignment.models import Assignment, AssignmentTargets
 from hx_lti_initializer.utils import debug_printer, get_lti_value, retrieve_token  # noqa
 from hx_lti_initializer.forms import CourseForm
-from hx_lti_initializer.utils import debug_printer, get_lti_value
+from hx_lti_initializer.utils import debug_printer, get_lti_value, save_session, create_new_user, initialize_lti_tool_provider, validate_request
 from django.conf import settings
 from abstract_base_classes.target_object_database_api import TOD_Implementation
 from django.contrib.sites.models import get_current_site
@@ -39,112 +38,6 @@ import requests
 
 import logging
 logging.basicConfig()
-
-
-# If we were to restructure (not recommended because then we can't reconcile with HX),
-# these first 4 functions would be in utils.py
-def validate_request(req):
-    """
-    Validates the request in order to permit or deny access to the LTI tool.
-    """
-    # print out the request to the terminal window if in debug mode
-    # this item is set in the settings, in the __init__.py file
-    if settings.LTI_DEBUG:
-        for item in sorted(req.POST.dict()):
-            debug_printer('DEBUG - %s: %s \r' % (item, req.POST[item]))
-
-    # verifies that request contains the information needed
-    if 'oauth_consumer_key' not in req.POST:
-        debug_printer('DEBUG - Consumer Key was not present in request.')
-        raise PermissionDenied()
-    if 'user_id' not in req.POST:
-        debug_printer('DEBUG - Anonymous ID was not present in request.')
-        raise PermissionDenied()
-    if ('lis_person_sourcedid' not in req.POST and
-            'lis_person_name_full' not in req.POST):
-        debug_printer('DEBUG - Username or Name was not present in request.')
-        raise PermissionDenied()
-
-
-def initialize_lti_tool_provider(req):
-    """
-    Starts the provider given the consumer_key and secret.
-    """
-    consumer_key = settings.CONSUMER_KEY
-    secret = settings.LTI_SECRET
-
-    # use the function from ims_lti_py app to verify and initialize tool
-    provider = DjangoToolProvider(consumer_key, secret, req.POST)
-
-    # NOTE: before validating the request, temporarily remove the
-    # QUERY_STRING to work around an issue with how Canvas signs requests
-    # that contain GET parameters. Before Canvas launches the tool, it duplicates the GET
-    # parameters as POST parameters, and signs the POST parameters (*not* the GET parameters).
-    # However, the oauth2 library that validates the request generates
-    # the oauth signature based on the combination of POST+GET parameters together,
-    # resulting in a signature mismatch. By removing the QUERY_STRING before
-    # validating the request, the library will generate the signature based only on
-    # the POST parameters like Canvas.
-    qs = req.META.pop('QUERY_STRING', '')
-
-    # now validate the tool via the valid_request function
-    # this means that request was well formed but invalid
-    if provider.valid_request(req) == False:
-        debug_printer("DEBUG - LTI Exception: Not a valid request.")
-        raise PermissionDenied()
-    else:
-        debug_printer('DEBUG - LTI Tool Provider was valid.')
-
-    req.META['QUERY_STRING'] = qs  # restore the query string
-
-    return provider
-
-
-def create_new_user(username, user_id, roles):
-    # now create the user and LTIProfile with the above information
-    # Max 30 length for person's name, do we want to change this? It's valid for HX but not ATG/FAS
-    try:
-        user = User.objects.create_user(username, user_id)
-    except IntegrityError:
-        # TODO: modify db to make student name not the primary key
-        # a temporary workaround for key integrity errors, until we can make the username not the primary key.
-        return create_new_user(username + " ", user_id, roles)
-    user.set_unusable_password()
-    user.is_superuser = False
-    user.is_staff = False
-
-    for admin_role in settings.ADMIN_ROLES:
-        for user_role in roles:
-            if admin_role.lower() == user_role.lower():
-                user.is_superuser = True
-                user.is_staff = True
-    user.save()
-    debug_printer('DEBUG - User was just created')
-
-    # pull the profile automatically created once the user was above
-    lti_profile = LTIProfile.objects.get(user=user)
-
-    lti_profile.anon_id = user_id
-    lti_profile.roles = (",").join(roles)
-    lti_profile.save()
-
-    return user, lti_profile
-
-
-def save_session(req, user_id, col_id, object_id, context_id, roles, is_staff):
-    if user_id is not None:
-        req.session["hx_user_id"] = user_id
-    if col_id is not None:
-        req.session["hx_collection_id"] = col_id
-    if object_id is not None:
-        req.session["hx_object_id"] = object_id
-    if context_id is not None:
-        req.session["hx_context_id"] = context_id
-    if roles is not None:
-        req.session["hx_roles"] = roles
-    if is_staff is not None:
-        req.session["is_staff"] = is_staff
-
 
 @csrf_exempt
 def launch_lti(request):
